% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canoodle_tbl.R
\name{canoodle_tbl}
\alias{canoodle_tbl}
\title{Create a tbl pointer to Canoodle.}
\usage{
canoodle_tbl(conn, name = NA)
}
\arguments{
\item{conn}{a DBI-compliant connection object. recommend using the 'connect_cdw()` function}

\item{name}{A character string specifying a table in the MY_DB.STG_CANOODLE schema}
}
\value{
a tbl or a display of tables and views in the STG_CANOODLE schema
}
\description{
If a name is not provided, will print a list of the available views in MY_DB.STG_CANOODLE.
}
\examples{

\dontrun{
conn <- connect_cdw()

canoodle_tbl(conn)

canoodle_tbl(conn, "ICU_STAY_DATAMART")

disconnect_cdw(conn)
}
}
