% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_cdw.R
\name{connect_cdw}
\alias{connect_cdw}
\title{Set up  a connection to Cloud Data Warehouse (CDW)}
\usage{
connect_cdw(dsn = "CDW", quiet = TRUE)
}
\arguments{
\item{quiet}{Default FALSE. if TRUE, then hide the display of the role, database, default table, and warehouse}

\item{database}{a valid database that the specified role can access. Default is PHC_DB_DEV}

\item{schema}{a valid schema that the specified role can access. Default is WHS_CORE_CDM}
}
\description{
ie. Snowflake, Synapse, etc.
}
\examples{

\dontrun{
conn <- connect_cdw()

ccdm_tbl(conn)

ccdm_tbl(conn, "HOSPITAL_ENCOUNTERS")

disconnect_cdw(conn)
}
}
