% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_CoreCDM.R
\name{create_CoreCDM}
\alias{create_CoreCDM}
\title{create_CoreCDM.R}
\usage{
create_CoreCDM(conn)
}
\arguments{
\item{conn}{a DBI-compliant connection object}
}
\description{
creates the main CoreCDM tables
}
