% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{get the results of a SQL query}
\usage{
read_sql(file_name)
}
\arguments{
\item{file_name}{character string with the path and filename of the .sql file}
}
\value{
concise sql statement ready to be passed to SQL Server/Snowflake, etc.
}
\description{
Reads an .sql file and prepares it for use with dbplyr/odbc queries to SQL Server/Snowflake, etc.
}
\details{
The original version of this function was developed by Thomas Huang.
}
\examples{

\dontrun{
sql_statement <- read_sql(here::here("SQL/demo.sql"))

# use dbGetQuery for single SQL clauses
results <- ccdm_query(sql_statement)

# use snowsql for complex SQL statements with multiple clauses
results2 <- system2("~/bin/snowsql", glue("-f {sql_statement} -o quiet={quiet}"))

}
}
