% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_CoreCDM.R
\name{update_CoreCDM}
\alias{update_CoreCDM}
\title{update_CoreCDM.R}
\usage{
update_CoreCDM(conn)
}
\arguments{
\item{conn}{a DBI-compliant connection object}
}
\description{
updates the CoreCDM tables with new/changed data
}
